Use openMINDS property names in almost all cases.
Previously, we used "name" instead of "full_name", "alias" instead of "short_name", and a couple of other substitutions
for backwards compatibility with pre-openMINDS schemas.

Since these variant names are very widely used, this commit also introduces aliases, so that the old names will still work.
The old names could be deprecated and phased out, or we could keep the aliases mechanism, as it might be useful in transitioning
between major versions of openMINDS.