A better way to make openMINDS terminology instances available as fairgraph instances.
The previous approach of initialising the instances as each module was imported, implemented in [b25091a5] and [e4df6ff0], has the problem that it prematurely triggers the resolution of types (in Property definitions) from strings to classes, and so the properties retain the openMINDS classes rather than getting the fairgraph ones.

The new approach leaves initialisation of instances until after all modules have been imported and the registry correctly set up.

This also reverts [5853d18f], since those changes are not needed if all classes are properly initialised.